refactor: split agent-loop into core and openai crates
- Rename everruns-agent-loop to everruns-core
  - Core abstractions for agent loop, events, tools, LLM providers
  - LLM provider traits use OpenAI Protocol as base
  - Includes capabilities system and built-in tools

- Create everruns-openai crate
  - OpenAI-specific LLM provider implementation
  - Implements core LlmProvider trait
  - Streaming and non-streaming support
  - Native types for OpenAI API

- Update everruns-worker
  - Remove providers module (moved to everruns-openai)
  - Use OpenAiProvider from everruns-openai
  - Use core types (LlmMessage, LlmCallConfig) instead of OpenAI types
  - Simplify adapters by removing OpenAiLlmAdapter wrapper

- Update specs/architecture.md with new crate structure

This refactoring provides cleaner separation between:
- Core abstractions (provider-agnostic traits and types)
- OpenAI implementation (provider-specific code)
