+++ crates/everruns-core/src/memory.rs
